Filtros

+++ Atividades/AulaEntityFramework/AulaEntityFramework/Repositories/IPessoaRepository.cs
@@ -5,7 +5,7 @@ namespace AulaEntityFramework.Repositories
 	public interface IPessoaRepository
 	{
 		public List<Pessoa>? GetAll();
-		public Pessoa? Get(int id);
+		public Pessoa? Get(long id);
 		public List<Pessoa>? GetByName(string? name);
 		public List<Pessoa>? GetByBirthDate(DateTime date);
 		public List<Pessoa>? GetByPeriodBirthDate(DateTime startDate, DateTime endDate);
@@ -13,6 +13,6 @@ namespace AulaEntityFramework.Repositories
 		public List<Pessoa>? GetByBirthMonth(int month);
 		public Pessoa Insert(Pessoa person);
 		public Pessoa Update(Pessoa person);
-		public Pessoa Delete(int id);
+		public Pessoa Delete(long id);
 	}
 }
